pyproject.toml
[tool.poetry]
name = "tap-jira"
version = "0.0.1"
description = "`tap-jira` is a Singer tap for tap-jira, built with the Meltano Singer SDK."
readme = "README.md"
authors = ["Ethan Stein"]
keywords = [
    "ELT",
    "tap-jira",
]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "<3.12,>=3.7.1"
singer-sdk = "^0.34.1"
fs-s3fs = { version = "^1.1.1", optional = true }
requests = "^2.28.2"
google-api-python-client= "^2.113.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
singer-sdk = { version="*", extras = ["testing"] }

[tool.poetry.extras]
s3 = ["fs-s3fs"]

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true

[tool.ruff]
ignore = [
    "ANN101",  # missing-type-self
    "ANN102",  # missing-type-cls
]
select = ["ALL"]
src = ["tap_jira"]
target-version = "py37"


[tool.ruff.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.isort]
known-first-party = ["tap_jira"]

[tool.ruff.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-jira = 'tap_jira.tap:TapJira.cli'
